Replace CmdAuthenticate::getUserObj() with AuthorizationManager::getPrivilegeDocument().
The new getPrivilegeDocument() returns a Status, and does not have a "pwd" out
parameter.  More importantly, by moving the implementation into the
AuthExternalState interface, we eliminate a symbol conflict between mongod and
mongos.

This implementation reduces code duplication, and meets the needs of the
enterprise SASL authentication module.  Once there is a mongod/mongos-agnostic
way to obtain a connection-like object for issuing queries, the code may be
futher simplified.

Conflicts:
	src/mongo/db/security.cpp